rs500_reader.py
import hid
import time
import paho.mqtt.client as mqtt
import os
import json

# Hier die Vendor ID (VID) und Product ID (PID) der RS500-Station eintragen
# Die Werte finden Sie mit 'lsusb' unter Linux
VENDOR_ID = 0x1234
PRODUCT_ID = 0x5678

# MQTT-Konfiguration aus den Umgebungsvariablen lesen
MQTT_HOST = os.environ.get('MQTT_HOST')
MQTT_USER = os.environ.get('MQTT_USER')
MQTT_PASSWORD = os.environ.get('MQTT_PASSWORD')

# MQTT-Client initialisieren
client = mqtt.Client()
client.username_pw_set(MQTT_USER, MQTT_PASSWORD)

# --- Die hier aufgeführten Funktionen sind Konzepte ---
# Sie müssen die Logik aus der von Ihnen bereitgestellten Anleitung hier umsetzen.
def get_device():
    # Findet das HID-Gerät anhand von VID und PID
    # Implementieren Sie hier die Logik, um das Gerät zu finden
    for d in hid.enumerate():
        if d['vendor_id'] == VENDOR_ID and d['product_id'] == PRODUCT_ID:
            print(f"Gerät gefunden: {d['path']}")
            return hid.Device(d['path'])
    return None

def read_data(device):
    # Sendet die HID-Anfragen und liest die Daten
    # Aus der Anleitung: 7 requests von 64 bytes
    # Implementieren Sie hier die Sende- und Empfangslogik
    # ... Beispielhaft
    request = [0x00, 0x01, 0x02, ...] # Daten aus der Anleitung
    device.write(request)
    data = device.read(64)
    return data

def parse_data(data):
    # Dekodiert die rohen HID-Daten in Temperatur und Luftfeuchtigkeit
    # Wie in der Anleitung beschrieben:
    # Temperatur ist ein 2-Byte-Integer, Luftfeuchtigkeit ein 1-Byte-Wert
    temp = ... # Ihre Dekodierungslogik
    hum = ... # Ihre Dekodierungslogik
    return temp, hum

def publish_to_mqtt(topic, payload):
    try:
        client.publish(topic, payload)
        print(f"Veröffentlicht auf {topic}: {payload}")
    except Exception as e:
        print(f"Fehler beim Veröffentlichen auf MQTT: {e}")

# --- Hauptschleife des Add-ons ---
if __name__ == "__main__":
    client.connect(MQTT_HOST, 1883, 60)
    client.loop_start()

    device = get_device()
    if not device:
        print("Gerät nicht gefunden. Wird in 60s erneut versucht.")
        time.sleep(60)
        exit(1)

    while True:
        try:
            raw_data = read_data(device)
            if raw_data:
                temp, hum = parse_data(raw_data)

                # Senden der Daten
                data_payload = {
                    "temperature": temp,
                    "humidity": hum
                }
                # Die Topics können frei gewählt werden
                # z.B. HomeAssistant Discovery
                publish_to_mqtt("homeassistant/sensor/rs500/main", json.dumps(data_payload))

        except Exception as e:
            print(f"Fehler im Lesevorgang: {e}. Versuche in 60s neu.")
            device.close()
            device = get_device()

        time.sleep(60) # Alle 60 Sekunden aktualisieren
