requirements.txt
hid
paho-mqtt
